extract Flag into separate file
